+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include congress.py
+include congress/*
 include test.py
 include requirements.txt
 include README.*
